msm: mdss: suppress dsi fifo errors during refresh rate update
When the frame rate is changed, there could be fifo errors for atleast
one frame and it logs in kernel messages. During the stability tests,
where there is continuous update to the frame rate, these error logs are
causes crash. This change suppresses the log when there is dynamic
refresh rate update.

Change-Id: I59020948fa33b326d691d24fe30ddcbde74b1816
Signed-off-by: Jeevan Shriram <jshriram@codeaurora.org>
Signed-off-by: Nirmal Abraham <nabrah@codeaurora.org>
Signed-off-by: Dede Dindin Qudsy <xtrymind@gmail.com>